Reorganize everything.
