+++ gulp/src/common/actions/error-actions.js
