+++ gulp/src/common/actions/loading-actions.js
