+++ gulp/src/common/reducers/error-reducer.js
