+++ gulp/src/common/reducers/loading-reducer.js
