+++ gulp/src/reporting/components/ReportingRouter.jsx
@@ -1,11 +1,11 @@
 import React, {Component, PropTypes} from 'react';
 import {Router, Route, IndexRedirect} from 'react-router';
-import DynamicReportApp from 'reporting/DynamicReportApp';
-import SetUpReport from 'reporting/SetUpReport';
-import ExportReport from 'reporting/ExportReport';
-import OldPreviewEmbedPage from 'reporting/OldPreviewEmbedPage';
+import DynamicReportApp from './DynamicReportApp';
+import SetUpReport from './SetUpReport';
+import ExportReport from './ExportReport';
+import OldPreviewEmbedPage from './OldPreviewEmbedPage';
 
-export class WizardRouter extends Component {
+export default class ReportingRouter extends Component {
   createElement(TheComponent, componentProps) {
     const {reportFinder} = this.props;
     const newProps = {...componentProps, reportFinder};
@@ -33,6 +33,6 @@ export class WizardRouter extends Component {
   }
 }
 
-WizardRouter.propTypes = {
+ReportingRouter.propTypes = {
   reportFinder: PropTypes.object.isRequired,
 };
