+++ gulp/src/reporting/reducers/report-list-reducer.js
