+++ gulp/src/reporting/reducers/report-state-reducer.js
